Generalized EconLit Scraper working
A generalized EconLit .xml scraper is working. `jel_scopus_matching.ipynb` is a notebook that a has a "MASTER CELL" that executes the scraping and organization based on a series of 'instruction' dictionaries, which it loops over.

TO DO:
Need to delete old per-journal .xml scrapers
Update README with how to access EconLit .xmls