Creating custom matching algorithm
00_Scraping/scopus_econlit_matching.iypynb
This script matches the database of papers collected from EconLit (which comes with JEL codes) with the database of papers collected from Scopus (in order to make sure that dataset is as complete as possible). It uses a custom nearest-neighbor matching algorithm for those observations in both sets that cannot be matched using only Volume, Issue Number and a naive (all-caps coercion) of the title. 

This algorithm generates a n x m matrix of scores that use various characteristics to identify a best-possible match. This commit is the beginning of this matching algorithm