Example function written
scopus_read.ipynb
We are going to use the same idea to modify the field_population() function to handle multiple authors who share an affiliation. It relies on an optional argument passed to field_population() which modifies the navigation_dict patterns to access different parts of the API-based JSON object.